Add files via upload
This script reads in the csv files provided by the competition hosts, wrangles the data, joins it into a table and creates pca plots to visualize the effects of our minority oversampling/ majority undersampling , as well as plots displaying the resulting balanced accuracy for the different techniques we tried. It also performs a grid search to optimize hyperparameters for the random forests used.